work done till refactoring the code

+++ 01-introduction-app/src/com/techlabs/tokens/Additions.java
@@ -6,7 +6,7 @@ public class Additions {
 		
 		byte number1=10;
 		byte number2=20;
-		byte sum=number1+number2;
+		byte sum=(byte)(number1+number2);
 		System.out.println("Addition of "+number1+" and "+number2+" is "+sum);
 
 	}
